rename reporting portfolio summary router
